Added Nginx proxy for HTTP log endpoints
In order to allow for secure transport of logs from Heroku it is
necessary to place an HTTPS proxy on the aggregator servers. This pull
request adds a `reverse_proxy` state file to install and configure Nginx
for this purpose. Also supported is specifying a token to be passed in
the query param in order to reject requests not containing the token.