Don't use git installation of Salt
The problematic Salt release has been withdrawn, so we can use the
"latest" one from their package, instead of installing with git.

A side effect of this is that we have to drop down to Debian 9 for the
"debian" box. As we do that, we have to change the package name of
libncurses-dev to libncurses5-dev.