Make compatible with Python 3
Update the FluentD config file Jinja2 template so that it works with
Python 3 as well as Python 2, and update the list of o/s package
dependencies.

Update the Vagrantfile to have Salt run under Python 3.

The changes to the Jinja2 template are to adapt to changes in the
OrderedDict class with Python 3. The template still works with Python 2.