Do not use Rbenv
Take out Rbenv, which was added in this PR. Use the system Ruby.

The only problem with this is that we have to comment out the CentOS VM,
because getting fluentd to work there is just too problematic. There is
no Salt package for CentOS 8, and CentOS 7 has too old a version of Ruby
(2.0.0) for fluentd itself and a number of its plugins. I was not able
to get fluentd to start after backing way down to the 0.12 release and
disabling all plugins. Our time will be better spent waiting for
SaltStack to release a package for CentOS 8.